Be more rigorous about ending the paxos phase when we have a ring
Previously we didn't cancel the paxosTicker when we heard about a ring.
And if we received a paxos-related gossip, we would pay attention to it
even if we already had a ring.

It's better and safer if we end all paxos-related activity when we get
a ring, if only to avoid unnecessary gossips.